Log aspect ratio buckets from SDXL pipelines.
