+++ src/invoke_training/training/pipelines/stable_diffusion/textual_inversion_sd.py
@@ -23,6 +23,7 @@ from invoke_training.training._shared.checkpoints.serialization import load_stat
 from invoke_training.training._shared.data.data_loaders.textual_inversion_sd_dataloader import (
     build_textual_inversion_sd_dataloader,
 )
+from invoke_training.training._shared.data.samplers.aspect_ratio_bucket_batch_sampler import log_aspect_ratio_buckets
 from invoke_training.training._shared.optimizer.optimizer_utils import initialize_optimizer
 from invoke_training.training._shared.stable_diffusion.model_loading_utils import load_models_sd
 from invoke_training.training._shared.stable_diffusion.textual_inversion import (
@@ -32,11 +33,7 @@ from invoke_training.training._shared.stable_diffusion.textual_inversion import 
     restore_original_embeddings,
 )
 from invoke_training.training._shared.stable_diffusion.validation import generate_validation_images_sd
-from invoke_training.training.pipelines.stable_diffusion.finetune_lora_sd import (
-    cache_vae_outputs,
-    log_aspect_ratio_buckets,
-    train_forward,
-)
+from invoke_training.training.pipelines.stable_diffusion.finetune_lora_sd import cache_vae_outputs, train_forward
 
 
 def _save_ti_embeddings(
